+++ src/main/kotlin/io/runescript/plugin/ide/spellchecker/RsBundledDictionaryProvider.kt
@@ -1,0 +1,7 @@
+package io.runescript.plugin.ide.spellchecker
+
+import com.intellij.spellchecker.BundledDictionaryProvider
+
+class RsBundledDictionaryProvider : BundledDictionaryProvider {
+    override fun getBundledDictionaries() = arrayOf("runescript.dic")
+}